requirements.txt
Appium-Python-Client==4.3.0
selenium==4.26.1
